requirements.txt
aiogram>=3.10
requests
beautifulsoup4
apscheduler
python-dotenv
